code linting XIII

+++ src/web/api_handlers_recordings_list.c
@@ -262,7 +262,7 @@ void handle_get_recordings(const http_request_t *req, http_response_t *res) {
         // Format file size for display (e.g., "1.8 MB")
         char size_str[32] = {0};
         if (recordings[i].size_bytes < 1024) {
-            snprintf(size_str, sizeof(size_str), "%ld B", recordings[i].size_bytes);
+            snprintf(size_str, sizeof(size_str), "%lu B", (unsigned long)recordings[i].size_bytes);
         } else if (recordings[i].size_bytes < (uint64_t)1024 * 1024) {
             snprintf(size_str, sizeof(size_str), "%.1f KB", (double)recordings[i].size_bytes / 1024.0);
         } else if (recordings[i].size_bytes < (uint64_t)1024 * 1024 * 1024) {
